pyproject.toml
[project]
name = "eunomia-ai"
version = "0.3.3"
description = "Authorization layer for AI Agents"
authors = [
    { name="What About You team", email="info@whataboutyou.ai" }
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "eunomia-core>=0.3.3",
    "fastapi[standard]>=0.115.9",
    "httpx>=0.28.1",
    "pydantic-settings>=2.8.1",
    "sqlalchemy>=2.0.38",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.9.10",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.5",
    "mkdocstrings[python]>=0.29.0",
]

[project.scripts]
eunomia = "eunomia.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eunomia"]

[tool.uv.workspace]
members = ["pkgs/core", "pkgs/sdks/python", "pkgs/extensions/*"]

[tool.uv.sources]
eunomia-core = { workspace = true }
